pyproject.toml
[project]
name = "vietcare-assistant"
version = "2.0.0"
description = "AI-powered Healthcare Q&A with RAG (Retrieval-Augmented Generation)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Duy Nguyen", email = "duynvt.work@gmail.com" }]

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[argon2]>=1.7",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.6",
    "openai>=1.0",
    "langchain>=0.2",
    "langchain-community>=0.2",
    "langchain-huggingface>=0.0.3",
    "faiss-cpu>=1.7",
    "sentence-transformers>=2.2",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "httpx>=0.27",
    "ruff>=0.3",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
